Add UserGroup permissions and API behaviour

+++ src/app/Services/UserGroupService.php
@@ -6,6 +6,7 @@ use App\Enums\User\UserGroup\UserGroupPrivacyEnum;
 use App\Http\Requests\User\UserGroup\UserGroupStoreRequest;
 use App\Http\Requests\User\UserGroup\UserGroupUpdateRequest;
 use App\Models\UserGroup;
+use Illuminate\Support\Facades\Gate;
 
 class UserGroupService
 {
@@ -19,6 +20,8 @@ class UserGroupService
 
     public function store(UserGroupStoreRequest $request): UserGroup
     {
+        Gate::authorize('create', UserGroup::class);
+
         $validated = $request->validated();
 
         $userGroup = UserGroup::create([
@@ -33,7 +36,10 @@ class UserGroupService
 
     public function show(string $id): UserGroup
     {
-        return UserGroup::findOrFail($id);
+        $userGroup = UserGroup::findOrFail($id);
+        Gate::authorize('view', $userGroup);
+
+        return $userGroup;
     }
 
     public function update(UserGroupUpdateRequest $request, string $id): UserGroup
@@ -41,6 +47,7 @@ class UserGroupService
         $validated = $request->validated();
         $userGroup = $this->show($id);
 
+        Gate::authorize('update', $userGroup);
         $userGroup->update($validated);
 
         return $userGroup;
@@ -49,6 +56,8 @@ class UserGroupService
     public function delete(string $id): void
     {
         $userGroup = UserGroup::findOrFail($id);
+
+        Gate::authorize('delete', $userGroup);
 
         $userGroup->delete();
     }
